move navigation types
